requirements.txt
azure-core
azure-search-documents
Flask==3.0.3
halo==0.0.31
langchain_community
langchain==0.1.16
numpy==1.24.4
openai==1.23.6
pandas==2.2.2
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.4.1.post1
scipy==1.13.0
tabulate==0.9.0
tqdm==4.66.2
google-auth
vertexai